put Climate change scripts on the server and linked the urls
