+++ testsite/polls/urls.py
@@ -8,4 +8,5 @@ urlpatterns = [
 	path( 'Slides.html/', views.slides, name='slides' ), 
 	path( 'index.html/', views.index, name='index' ),
 	path( 'bio_output/', views.biodiversity_submit, name='bio_submit' ),
+	path( 'climate_output/', views.climate_submit, name='climate_submit' ), 
 ]
