+++ testsite/templates/Slides.html
@@ -167,6 +167,7 @@ window.onclick = function(event) {
 	{{message}}
 	{% endif %}
 
+
 <form action="">
   Longitude:<br>
   <input type="text" name="Longitude" value="">
@@ -174,8 +175,11 @@ window.onclick = function(event) {
   Latitude:<br>
   <input type="text" name="Latitude" value="">
   <br><br>
-  <input type="submit" value="Submit">
 </form>
+<button type='submit' onclick="location.href='{% url 'climate_submit' %}'">climate submit</button><hr>
+	{% if message %}
+	{{message}}
+	{% endif %}
